sampler/indexer: Don't use 32-bit pointIndex.
I could not find any evidence why there should be less than 32-bit many,
so use 64-bit.

+++ Converter/include/sampler_random.h
@@ -21,7 +21,7 @@ struct SamplerRandom : public Sampler {
 			double x;
 			double y;
 			double z;
-			int32_t pointIndex;
+			int64_t pointIndex;
 			int32_t childIndex;
 		};
